Timer update
updated [timer] so that it responds properly no matter how global
transport is started. Before it used to work properly only if started
by Push. That required moving from n to ms. In this way the timer is
always banging. I also updated metro speed to 33ms. Previous speed of
64n was unnecessary high.